Fixed fireball not being destroyed when hitting non-Enemy object
I noticed that the fireball was only getting destroyed when it hit an enemy because I moved the Destroy() call to the ragdoll coroutine. Reorganized the OnTriggerEnter() function to instantiate explosion first and then added an else to the following "if tag==Enemy" block to still Destroy the game object if it collides with anything else.