Implemented Camera Shake and reorganized Asset folder
Reorganized the asset folder so that all of our main folders start with an underscore, placing them at the top of the Asset folder.

Set up code for camera shake and tested it out by implementing a shake for the fireball explosion.

To create a new shake event, right-click in the asset folder and select Create > Custom > Camera Shake Event.